pyproject.toml
[tool.poetry]
name = "powerdict"
version = "1.0.11"
description = "A power station dictionary that enables mapping between various naming conventions and associated plant metadata"
authors = ["AyrtonB <ayrtonbourn@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^1.5.3"
matplotlib = "^3.7.1"
seaborn = "^0.12.2"
mkdocs = "^1.4.2"
mkdocs-material = "^9.1.2"
typer = "^0.7.0"
tqdm = "^4.65.0"
Jinja2 = "^3.1.2"
mkdocs-material-extensions = "^1.1.1"
sqlmodel = "^0.0.8"
email-validator = "^1.3.1"
loguru = "^0.6.0"
fastapi = "^0.95.0"
uvicorn = "^0.21.1"
itsdangerous = "^2.1.2"
python-jose = "^3.3.0"
passlib = "^1.7.4"
python-multipart = "^0.0.6"
bcrypt = "^4.0.1"
httpx = "^0.24.0"
ipykernel = "^6.22.0"
lxml = "^4.9.2"
html5lib = "^1.1"
aiohttp = "^3.8.4"

[tool.poetry.dev-dependencies]
ipykernel = "^6.19.2"
python-dotenv = "^0.21.0"
black = "22.3.0"
isort = "5.10.1"
mypy = "0.971"
pre-commit = "2.13.0"
pytest = "7.1.3"
jupyterlab = "^3.6.3"

[tool.isort]
multi_line_output = 3
line_length = 88
include_trailing_comma = true
force_grid_wrap = 0
combine_as_imports = true

[tool.black]
line-length = 120
skip-string-normalization = true

[tool.mypy]
strict = true

[build-system]
requires = ["poetry-core>=1.5.0"]
build-backend = "poetry.core.masonry.api"
